requirements/base.txt
tortoise-orm==0.16.13
aiogram==2.9.2